US-3382, US-3381 Added some comments and descriptions for the test classes
